#122: Renamed ParseValueListTest to ImmutableListTest.
